Step_3/requirements.txt
# Default Dependencies
docling==2.31.0
google-generativeai==0.8.3
sentence-transformers>=2.2.0
lancedb==0.22.0
python-dotenv>=0.21.0
pydantic>=2.0.0

# Step 3: Advanced RAG Dependencies
torch>=1.9.0
transformers>=4.20.0
rank-bm25>=0.2.2
scikit-learn>=1.0.0
nltk>=3.7
rouge-score>=0.1.2
evaluate>=0.4.0
numpy>=1.21.0
pandas>=1.3.0

